// src/app/modules/digital-factory/Models/df-supplier-directory-master-dto.ts
export interface DfSupplierDirectoryMasterDto {
  supplierId: number;
  vendorName?: string;
  missionStatement?: string;
  countryId: number;
  regionId: number;
  state?: string;
  address?: string;
  city?: string;
  zipCode?: string;
  phone?: string;
  owner?: string;
  contactEmail?: string;
  products?: string;
  website?: string;
  companyType: number;
  diverse?: number;
  qualityCertifications?: string;
  registrations?: string;
  companySizeBySales?: string;
  companySizeByEmployees?: string;
  yearFounded?: string;
  dunsId?: string;
  eid?: string;
  financialRisk: number;
  importExportClassifications?: string;
  parentCompany?: string;
  partners?: string;
  carbonNeutral?: number;
  preferredUOM?: number;
  interestRate?: number;
  profitMargin?: number;
  materialProfitMargin?: number;
  manufacturingProfitMargin?: number;
  materialOverhead?: number;
  factoryOverhead?: number;
  sgAndA?: number;
  carryingCostsForRawMaterial?: number;
  carryingCostsForFinishedGoods?: number;
  carryingCostsForPaymentTerms?: number;
  hoursPerShift?: number;
  shiftsPerDay?: number;
  workingDaysPerYear?: number;
  totalBreaksPerShift?: number;
  factoryTotalSize?: number;
  factoryType?: string;
  annualHours?: number;
  isActive?: boolean;
  countryName?: string;
  regionName?: string;
  laborSkilledCost?: number;
  laborSkilledWorkers?: number;
  laborLowSkilledCost?: number;
  laborLowSkilledWorkers?: number;
  laborSemiSkilledCost?: number;
  laborSemiSkilledWorkers?: number;
  laborSpecialSkilledCost?: number;
  laborSpecialSkilledWorkers?: number;
  laborQualityCost?: number;
  laborQualityWorkers?: number;
  laborProgrammerCost?: number;
  laborProgrammerWorkers?: number;
  laborToolDesignerCost?: number;
  laborToolDesignerWorkers?: number;
  laborToolMakerCost?: number;
  laborToolMakerWorkers?: number;
  coalCost?: number;
  windCost?: number;
  naturalGasCost?: number;
  nuclearCost?: number;
  geothermalCost?: number;
  otherNonRenewableCost?: number;
  otherRenewableCost?: number;
  coalPortion?: number;
  windPortion?: number;
  naturalGasPortion?: number;
  nuclearPortion?: number;
  geothermalPortion?: number;
  otherNonRenewablePortion?: number;
  otherRenewablePortion?: number;
  coalESG?: number;
  windESG?: number;
  naturalGasESG?: number;
  nuclearESG?: number;
  geothermalESG?: number;
  otherNonRenewableESG?: number;
  otherRenewableESG?: number;
  rentRate?: number;
  latitude?: number;
  longitude?: number;
  // labourRateMasterDto?: LaborRateMasterDto
}