// src/app/modules/digital-factory/Models/df-supplier-cost-structure-info.ts
export interface DfSupplierCostStructureInfo {
  interestRate?: number;
  profitMargin?: number;
  materialOverhead?: number;
  factoryOverhead?: number;
  sgAndA?: number;
  carryingCostsForRawMaterial?: number;
  carryingCostsForFinishedGoods?: number;
  carryingCostsForPaymentTerms?: number;
  rentRate?: number;
}